feat: Created controller for the admin page
Created controller for the admin page and created the PHP files for the pages under admin pages. Edited the routes to allow 2 input parameters for the controller and fixed some logic so entering the admin page will direct you to the login page and using the address bar to login wont be possible.